refactor(akai-fire): move fire-only value types out of framework namespace
