--- modules/akai-fire/src/main/java/com/bitwig/extensions/framework/values/KeyboardLayoutType.java
@@ -1,6 +1,0 @@
-package com.bitwig.extensions.framework.values;
-
-public enum KeyboardLayoutType {
-    ISOMORPHIC,
-    IMITATE_KEYS
-}
