+++ modules/akai-fire/src/main/java/com/oikoaudio/fire/note/NoteMode.java
@@ -20,7 +20,7 @@ import com.bitwig.extensions.framework.MusicalScale;
 import com.bitwig.extensions.framework.MusicalScaleLibrary;
 import com.bitwig.extensions.framework.values.BooleanValueObject;
 import com.bitwig.extensions.framework.values.Midi;
-import com.bitwig.extensions.framework.values.StepViewPosition;
+import com.oikoaudio.fire.values.StepViewPosition;
 import com.oikoaudio.fire.AkaiFireOikontrolExtension;
 import com.oikoaudio.fire.ColorLookup;
 import com.oikoaudio.fire.FireControlPreferences;
