+++ modules/akai-fire/src/main/java/com/oikoaudio/fire/values/KeyboardLayoutType.java
@@ -1,0 +1,6 @@
+package com.oikoaudio.fire.values;
+
+public enum KeyboardLayoutType {
+    ISOMORPHIC,
+    IMITATE_KEYS
+}
